Fix ruff RUF046: drop redundant int() around ceil() in etl_runtime_creep
math.ceil already returns int, so int(ceil(...)) is a redundant cast. No
behavior change (runs_to_2x stays an int).

+++ app/logic/insights.py
@@ -838,7 +838,7 @@ def etl_runtime_creep(
         projected = latest + slope * horizon_runs
         target = 2.0 * baseline
         # runs until the projection reaches 2× baseline; 0 = already there. slope > 0 here, so safe.
-        runs_to_2x = 0 if latest >= target else int(ceil((target - latest) / slope))
+        runs_to_2x = 0 if latest >= target else ceil((target - latest) / slope)
         rows.append({
             "WORKFLOW_NAME": wf, "TASK_NAME": task, "RUNS": n,
             "LATEST_SEC": round(latest, 1), "BASELINE_SEC": round(baseline, 1),
